uncomment momo's code

+++ PRAdjust.java
@@ -30,11 +30,11 @@ public class PRAdjust {
     public static class PRAdjustReducer extends Reducer<IntWritable, PDNodeWritable, IntWritable, PDNodeWritable> { 
         public void reduce(IntWritable key, PDNodeWritable node, Context context) throws IOException, InterruptedException {
             Configuration conf = context.getConfiguration();
-            double alpha = Long.parseLong(conf.get("alpha")); // Not yet set
+            double alpha = Double.parseDouble(conf.get("alpha"));
             long nodeCount = Long.parseLong(conf.get("nodeCount"));
             double m = context.getCounter( m ).getValue();
             double p = node.getPRValue().get();
-            double p2 = (alpha/nodeNum) + (1-alpha) * ((m/nodeNum)+p);
+            double p2 = (alpha / nodeNum) + (1.0 - alpha) * ((m / nodeNum) + p);
             node.setPRValue(new DoubleWritable(p2));
             context.write(key, node);
         }
